requirements.txt
requests==2.32.3
numpy==1.26.4
python-dotenv==1.0.1
fastapi[standard]==0.112.0
pillow==10.4.0